tests: update test
