+++ tests/pr_test.go
@@ -15,6 +15,7 @@ import (
 // Use existing resource group for tests
 const resourceGroup = "geretain-test-key-protect-all-inclusive"
 const solutionDADir = "solutions/standard"
+const advancedExampleTerraformDir = "examples/advanced"
 
 // Define a struct with fields that match the structure of the YAML data
 const yamlLocation = "../common-dev-assets/common-go-assets/common-permanent-resources.yaml"
@@ -88,4 +89,33 @@ func TestRunUpgradeDASolution(t *testing.T) {
 		assert.Nil(t, err, "This should not have errored")
 		assert.NotNil(t, output, "Expected some output")
 	}
+}
+
+func TestRunAdvanceExample(t *testing.T) {
+	t.Parallel()
+
+	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
+		Testing: t,
+		Prefix:  "kms-all-inc-advanced",
+		TarIncludePatterns: []string{
+			"*.tf",
+			advancedExampleTerraformDir + "/*.tf",
+		},
+
+		ResourceGroup:          resourceGroup,
+		TemplateFolder:         advancedExampleTerraformDir,
+		Tags:                   []string{"test-schematic"},
+		DeleteWorkspaceOnFail:  false,
+		WaitJobCompleteMinutes: 60,
+	})
+
+	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
+		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
+		{Name: "region", Value: options.Region, DataType: "string"},
+		{Name: "prefix", Value: options.Prefix, DataType: "string"},
+		{Name: "resource_group", Value: options.ResourceGroup, DataType: "string"},
+	}
+
+	err := options.RunSchematicTest()
+	assert.Nil(t, err, "This should not have errored")
 }
